# Add accessor for both time and count, and fix extra counts that overlap flip calls