- Disable a line that increments nCount for a frame timer that is active during a MicroProfileFlip. Now nCount is only incremented when a token exit (MP_LOG_LEAVE, due to calling MicroProfileLeave) is handled.
- Add MicroProfileGetTimeAndCount which returns time like MicroProfileGetTime, but also a uint32_t count and whether a token was found.

Before this change, a timing around a 5 second piece of code (at 60 fps) would show up (via MicroProfileGetTime and MicroProfileGetFrameCount) as 301 timings, which are ~16ms each.
After this change, that timing would show up as 1 5000ms event.

I checked the HTML view as well, and everything still seems to work fine both with or without this change. Zooming in on one frame in the middle of the above 5 second example would show ~16ms of time spent on 0 count, with NaN average time.

The thing motivating this change is that I optimized some code that happened frequently on a background thread, going from about 20 msec to 2 msec per call. The rate of calling that code didn't change, but the profiling stats of the 'before' code showed roughly twice as many call events as the 'after'. Since every 'before' call overlapped a MicroProfileFlip(), they were counted twice, and their average time was thus measured as roughly half what it should have been. After the change, both 'before' and 'after' profiling results had the same number of calls to the instrumented code.